fix: prettier config and replaced logo
