--- platform/web/.prettierrc.cjs
@@ -1,9 +1,0 @@
-module.exports = {
-	printWidth: 80,
-	tabWidth: 2,
-	useTabs: true,
-	singleQuote: false,
-	semi: true,
-	trailingComma: "none",
-	bracketSpacing: true
-};
