+++ platform/web/.prettierrc.json
@@ -1,0 +1,10 @@
+{
+	"$schema": "http://json.schemastore.org/prettierrc",
+	"printWidth": 80,
+	"tabWidth": 2,
+	"useTabs": true,
+	"singleQuote": false,
+	"semi": true,
+	"trailingComma": "none",
+	"bracketSpacing": true
+}
