+++ platform/web/index.html
@@ -2,10 +2,10 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <link rel="icon" type="image/svg+xml" href="/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <link rel="stylesheet" href="/path/to/reset-css/reset.css" />
-    <title>Vite + React + TS</title>
+    <title>一站式 RPC 调用平台</title>
   </head>
   <body>
     <div id="root"></div>
