+++ platform/web/src/App.module.scss
@@ -1,11 +1,6 @@
 .layout {
 	height: 100vh;
 	background-color: var(--semi-color-bg-1);
-	.header {
-		position: absolute;
-		width: 100vw;
-		z-index: 999;
-	}
 	.nav {
 		height: 6vh;
 	}
@@ -21,26 +16,6 @@
 				display: none;
 			}
 		}
-	}
-	.logo {
-		font-size: 2rem;
-		color: var(--semi-color-text-0);
-	}
-	.logo-title {
-		margin-right: 24px;
-		color: var(--semi-color-text-0);
-		font-weight: 600;
-	}
-	.footer {
-		display: flex;
-		justify-content: space-between;
-		padding: 20px;
-		color: var(--semi-color-text-2);
-		background-color: rgba(var(--semi-grey-0), 1);
-	}
-	.copyright {
-		display: flex;
-		align-items: center;
 	}
 	.sider {
 		background-color: var(--semi-color-bg-1);
