+++ platform/web/src/App.tsx
@@ -1,43 +1,37 @@
 import { Layout, Nav, Breadcrumb, Skeleton } from "@douyinfe/semi-ui";
-import {
-	IconCode,
-	IconBytedanceLogo,
-	IconHome,
-	IconHistogram,
-	IconLive
-} from "@douyinfe/semi-icons";
+import { IconCode, IconGitlabLogo, IconFolderOpen } from "@douyinfe/semi-icons";
 import { Outlet, useNavigate } from "react-router-dom";
 import styles from "./App.module.scss";
-import { SwitchButton } from "./components/SwitchButton";
 import { useEffect, useState } from "react";
+import Header from "./components/Header";
+import Footer from "./components/Footer";
 
 export default function App() {
-	const { Header, Footer, Sider, Content } = Layout;
+	const { Sider, Content } = Layout;
 	const [selectedKey, setSelectedKey] = useState("idl");
-	const navigate = useNavigate();
-	const themeMode = localStorage.getItem("theme-mode");
 	const navList = [
 		{
 			itemKey: "idl",
 			text: "IDL 信息查询",
-			icon: <IconHome size="large" />
+			icon: <IconCode size="large" />
 		},
 		{
 			itemKey: "repository",
 			text: "仓库信息管理",
-			icon: <IconHistogram size="large" />
+			icon: <IconGitlabLogo size="large" />
 		},
 		{
 			itemKey: "template",
 			text: "模版管理",
-			icon: <IconLive size="large" />
+			icon: <IconFolderOpen size="large" />
 		}
 	];
 
+	// 主题模式
+	const themeMode = localStorage.getItem("theme-mode");
 	if (themeMode) {
 		document.body.setAttribute("theme-mode", themeMode);
 	}
-	// 开启监听 storage 事件
 	window.addEventListener("storage", (event) => {
 		if (event.key === "theme-mode") {
 			const body = document.body;
@@ -49,23 +43,15 @@ export default function App() {
 		}
 	});
 
+	// 路由跳转
+	const navigate = useNavigate();
 	useEffect(() => {
 		navigate(selectedKey);
 	}, [navigate, selectedKey]);
 
 	return (
 		<Layout className={styles["layout"]}>
-			<Header className={styles["header"]}>
-				<Nav mode="horizontal" className={styles["nav"]}>
-					<Nav.Header>
-						<IconCode className={styles["logo"]} />
-					</Nav.Header>
-					<span className={styles["logo-title"]}>一站式 RPC 调用平台</span>
-					<Nav.Footer>
-						<SwitchButton />
-					</Nav.Footer>
-				</Nav>
-			</Header>
+			<Header />
 			<Layout className={styles["inner-layout"]}>
 				<Sider className={styles["sider"]}>
 					<Nav
@@ -103,15 +89,7 @@ export default function App() {
 					</Skeleton>
 				</Content>
 			</Layout>
-			<Footer className={styles["footer"]}>
-				<span className={styles["copyright"]}>
-					<IconBytedanceLogo size="large" style={{ marginRight: "8px" }} />
-					<span>Copyright © 2023. All Rights Reserved. </span>
-				</span>
-				<span>
-					<span>反馈建议</span>
-				</span>
-			</Footer>
+			<Footer />
 		</Layout>
 	);
 }
