+++ platform/web/src/components/Header/index.module.scss
@@ -1,0 +1,15 @@
+.header {
+    position: absolute;
+    width: 100vw;
+    z-index: 999;
+}
+.nav {
+    height: 6vh;
+}
+.logo {
+    height: 5vh;
+}
+.logo-title {
+    color: var(--semi-color-text-0);
+    font-weight: 600;
+}
