+++ platform/web/src/index.css
@@ -3,8 +3,8 @@
   padding: 0;
   box-sizing: border-box;
   font-family: 'Roboto', sans-serif;
-  font-size: 16px;
 }
+
 a {
   text-decoration: none;
   color: var(--semi-color-primary);
