+++ platform/web/src/main.tsx
@@ -1,39 +1,7 @@
 import ReactDOM from "react-dom/client";
-import {
-	createBrowserRouter,
-	RouterProvider,
-	Navigate
-} from "react-router-dom";
-import App from "./App.tsx";
+import { RouterProvider } from "react-router-dom";
+import router from "./routes/index.tsx";
 import "./index.css";
-import IdlPage from "./pages/IdlPage";
-import RepositoryPage from "./pages/RepositoryPage";
-import TemplatePage from "./pages/TemplatePage";
-
-const router = createBrowserRouter([
-	{
-		path: "/",
-		element: <App />,
-		children: [
-			{
-				path: "/idl",
-				element: <IdlPage />
-			},
-			{
-				path: "/repository",
-				element: <RepositoryPage />
-			},
-			{
-				path: "/template",
-				element: <TemplatePage />
-			},
-			{
-				path: "*",
-				element: <Navigate to={"/idl"} />
-			}
-		]
-	}
-]);
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
 	<RouterProvider router={router} />
